Pull module signing into its own module
